Correct return type for getSetOfEggCountsv2
Since we divide the number of eggs by a float,
the result is an array of floats not ints

+++ sch_simulation/helsim_FUNC_KK/utils.py
@@ -80,7 +80,7 @@ def getSetOfEggCountsv2(
     Unfertilized: bool,
     nSamples: int = 2,
     surveyType: str = 'KK2'
-) -> NDArray[np.int_]:
+) -> NDArray[np.float_]:
 
     """
     This function returns a set of readings of egg counts from a vector of individuals,
